+++ spec/skeem/datum_dsl_spec.rb
@@ -35,6 +35,15 @@ module Skeem
         ['+456', 456]
       ]
     end
+    
+    let(:rational_tests) do
+      [
+        [-Rational(2,3), -Rational(2, 3)],
+        [Rational(22, 7), Rational(22, 7)],
+        ['-2/3', -Rational(2, 3)],
+        ['+22/7', Rational(22, 7)]
+      ]
+    end    
 
     let(:real_tests) do
       [
@@ -83,6 +92,12 @@ module Skeem
           expect(subject.integer(literal)).to eq(predicted)
         end
       end
+      
+      it 'should convert rational literals' do
+        rational_tests.each do |(literal, predicted)|
+          expect(subject.rational(literal)).to eq(predicted)
+        end
+      end      
 
       it 'should convert real number literals' do
         real_tests.each do |(literal, predicted)|
@@ -163,6 +178,12 @@ module Skeem
           expect(subject.to_datum(literal)).to eq(predicted)
         end
       end
+      
+      it 'should recognize & convert rational literals' do
+        rational_tests.each do |(literal, predicted)|
+          expect(subject.to_datum(literal)).to eq(predicted)
+        end
+      end      
 
       it 'should recognize & convert real number literals' do
         real_tests.each do |(literal, predicted)|